Float-to-double cleanup: recipes, serializers, block entity, packets, compat; fix isEmpty() bug
Agent-Logs-Url: https://github.com/VincentHuto/Hemomancy/sessions/7c177e51-93a0-4759-90b7-7c118a87915a

+++ src/main/java/com/vincenthuto/hemomancy/client/screen/StructureSpawnerScreen.java
@@ -235,7 +235,7 @@ public class StructureSpawnerScreen extends AbstractContainerScreen<StructureSpa
 	}
 
 	private static int getRequiredDegreeForStructure(BloodStructureRecipe recipe) {
-		float cost = recipe.getBloodCost();
+		double cost = recipe.getBloodCost();
 		for (int i = 0; i < CRAFTING_TIER_THRESHOLDS.length; i++) {
 			if (cost <= CRAFTING_TIER_THRESHOLDS[i]) {
 				return CRAFTING_TIER_DEGREE_REQ[i];
